fix(gasprof): resolve lint findings and make filetest store-gas test cache-robust
Fixes the failing CI lint and one environment-fragile test on the gas
profiler PR.

- unconvert: drop no-op conversions (store.Gas is an int64 alias, so
  int64(...)/store.Gas(...) around already-int64 values are redundant) in
  gasprof.go, vm/keeper.go, and the gasprof test files.
- dogsled: TestGasProfile_filetestChargesStoreGas captured RunFiletest's
  report string instead of a third blank identifier, and now uses it as the
  require failure message.
- Make TestGasProfile_filetestChargesStoreGas cache-robust: a bare
  `package main` filetest only charges store READ gas (loading types),
  which a warm object cache can drive to zero, so it passed in isolation
  but failed under the full-package CI run. Switch to a realm filetest that
  mutates persistent state, forcing deterministic store WRITES whose gas
  cannot be cache-elided.

+++ gno.land/pkg/sdk/vm/gasprof_test.go
@@ -59,7 +59,7 @@ func Append(cur realm, n int) int {
 	// Reconciliation: the whole tree (ante + all dimensions minus refunds)
 	// equals the tx meter's total. Nothing dropped or double-counted.
 	net := tot.CPU + tot.Alloc + tot.Store + tot.Other - tot.Refund
-	require.Equal(t, int64(ctx.GasMeter().GasConsumed()), net,
+	require.Equal(t, ctx.GasMeter().GasConsumed(), net,
 		"profile reconciles with the tx meter")
 
 	// Attribution: the called function is a nested node (leading ';' proves it
@@ -112,7 +112,7 @@ func Append(cur realm, n int) int {
 		require.NoError(t, env.vmk.AddPackage(ctx, NewMsgAddPackage(addr, "gno.land/r/test", files)))
 		_, err := env.vmk.Call(ctx, NewMsgCall(addr, nil, "gno.land/r/test", "Append", []string{"40"}))
 		require.NoError(t, err)
-		return int64(ctx.GasMeter().GasConsumed())
+		return ctx.GasMeter().GasConsumed()
 	}
 	require.Equal(t, run(false), run(true), "profiling must not change gas consumed")
 }
@@ -153,7 +153,7 @@ func main() {
 	require.Positive(t, tot.CPU, "cpu captured across both Run machines")
 	require.Equal(t, int64(9_000), tot.Other, "(ante) snapshot booked")
 	net := tot.CPU + tot.Alloc + tot.Store + tot.Other - tot.Refund
-	require.Equal(t, int64(ctx.GasMeter().GasConsumed()), net, "Run profile reconciles with the tx meter")
+	require.Equal(t, ctx.GasMeter().GasConsumed(), net, "Run profile reconciles with the tx meter")
 
 	var fb strings.Builder
 	require.NoError(t, prof.WriteFolded(&fb))
